// src/types/IOmsAccountStateChangeSignal.ts
export interface IOmsAccountStateChangeSignal {
    totalAsset: number;
    buyingPower: number;
    blokedValue: number;
    investorBourseCodeId: number;
    investorBourseCodeIsin: number;
    creditBlockedMoney: number;
    creditMoneyRemain: number;
    creditMoney: number;
    nonWithdrawableBlockedMoney: number;
    nonWithdrawableMoneyRemain: number;
    withdrawableBlockedMoney: number;
    withdrawableMoneyRemain: number;
}